# micro-optimizations in ProtocolWriter