Resolves #319 

New MicroBenchmark shows benefit in writing UInt 16/32/64.  Maybe because I am on a newer CPU on this box?  :man_shrugging: Looks like it is worth changing though

```

BenchmarkDotNet v0.13.12, Ubuntu 22.04.3 LTS (Jammy Jellyfish)
AMD Ryzen 5 3600, 1 CPU, 12 logical and 6 physical cores
.NET SDK 8.0.101
  [Host]   : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX2
  ShortRun : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX2

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```
| Method            | Iter    | Size | Mean       | Error     | StdDev  | Allocated |
|------------------ |-------- |----- |-----------:|----------:|--------:|----------:|
| **PubCopy**           | **1000000** | **2**    |   **368.9 μs** | **109.60 μs** | **6.01 μs** |         **-** |
| PubSetIndex       | 1000000 | 2    |   485.2 μs |  59.42 μs | 3.26 μs |         - |
| PubBinaryConst    | 1000000 | 2    |   245.7 μs |  18.84 μs | 1.03 μs |         - |
| PubBinaryReadonly | 1000000 | 2    |   242.5 μs |  56.04 μs | 3.07 μs |         - |
| **PubCopy**           | **1000000** | **4**    |   **483.8 μs** |  **72.14 μs** | **3.95 μs** |         **-** |
| PubSetIndex       | 1000000 | 4    |   969.0 μs |  42.56 μs | 2.33 μs |       1 B |
| PubBinaryConst    | 1000000 | 4    |   247.9 μs |  94.50 μs | 5.18 μs |         - |
| PubBinaryReadonly | 1000000 | 4    |   244.6 μs |  19.72 μs | 1.08 μs |         - |
| **PubCopy**           | **1000000** | **8**    |   **495.1 μs** | **107.94 μs** | **5.92 μs** |         **-** |
| PubSetIndex       | 1000000 | 8    | 1,916.3 μs |   7.41 μs | 0.41 μs |       1 B |
| PubBinaryConst    | 1000000 | 8    |   366.2 μs |  87.33 μs | 4.79 μs |         - |
| PubBinaryReadonly | 1000000 | 8    |   368.6 μs | 128.24 μs | 7.03 μs |         - |


Also added Throw Helper to Protocol Writer as suggested in #320